chore: change attendanceCount from Integer to Long in PopularSessionResponse
- Update attendanceCount field type from Integer to Long
- Prepare for larger attendance counts in popular sessions analytics
- Maintain same builder pattern and lombok annotations

+++ src/main/java/com/dev/quikkkk/dto/response/PopularSessionResponse.java
@@ -14,5 +14,5 @@ import lombok.Setter;
 public class PopularSessionResponse {
     private String sessionId;
     private String trainerName;
-    private Integer attendanceCount;
+    private Long attendanceCount;
 }
